feat(symfony) Add LuckyController and corresponding view
This commit introduces a new controller, "LuckyController", and its corresponding view "number.html.twig". It includes a new route '/aloha/number' that returns a random number to be displayed in the view. The purpose of this addition is to provide a basic number generation functionality that can be used and expanded upon in the future. Also, the "composer.json" file and "composer.lock" files were updated to reflect recent changes in the project's dependencies, particularly the addition of "twig/twig".